manual-scans/requirements.txt
boto3==1.20.24
dnspython==2.1.0
requests==2.26.0